refactor: safer codeshare
