+++ AdminUi/apps/admin_ui/lib/core/widgets/widgets.dart
@@ -1,2 +1,3 @@
 export 'app_title.dart';
 export 'filters/filters.dart';
+export 'identities_data_table/identities_data_table.dart';
