+++ AdminUi/apps/admin_ui/lib/home/identities_overview/identities_overview.dart
@@ -1,10 +1,7 @@
 import 'package:admin_api_sdk/admin_api_sdk.dart';
-import 'package:data_table_2/data_table_2.dart';
 import 'package:flutter/material.dart';
 
 import '/core/core.dart';
-import 'identities_data_table_source.dart';
-import 'identities_filter.dart';
 
 class IdentitiesOverview extends StatefulWidget {
   const IdentitiesOverview({super.key});
@@ -15,11 +12,6 @@ class IdentitiesOverview extends StatefulWidget {
 
 class _IdentitiesOverviewState extends State<IdentitiesOverview> {
   late IdentityDataTableSource _dataSource;
-
-  int _sortColumnIndex = 0;
-  bool _sortColumnAscending = true;
-
-  int _rowsPerPage = 5;
 
   @override
   void didChangeDependencies() {
@@ -51,60 +43,11 @@ class _IdentitiesOverviewState extends State<IdentitiesOverview> {
                     ..refreshDatasource();
                 },
               ),
-              Expanded(
-                child: AsyncPaginatedDataTable2(
-                  rowsPerPage: _rowsPerPage,
-                  onRowsPerPageChanged: _setRowsPerPage,
-                  sortColumnIndex: _sortColumnIndex,
-                  sortAscending: _sortColumnAscending,
-                  showFirstLastButtons: true,
-                  columnSpacing: 5,
-                  source: _dataSource,
-                  isVerticalScrollBarVisible: true,
-                  renderEmptyRowsInTheEnd: false,
-                  availableRowsPerPage: const [5, 10, 25, 50, 100],
-                  empty: const Text('No identities found.'),
-                  errorBuilder: (error) => Center(
-                    child: Column(
-                      mainAxisSize: MainAxisSize.min,
-                      children: [
-                        const Text('An error occurred loading the data.'),
-                        Gaps.h16,
-                        FilledButton(onPressed: () => _dataSource.refreshDatasource(), child: const Text('Retry')),
-                      ],
-                    ),
-                  ),
-                  columns: <DataColumn2>[
-                    DataColumn2(label: const Text('Address'), size: ColumnSize.L, onSort: _sort),
-                    const DataColumn2(label: Text('Tier'), size: ColumnSize.S),
-                    DataColumn2(label: const Text('Created with Client'), onSort: _sort),
-                    DataColumn2(label: const Text('Number of Devices'), onSort: _sort),
-                    DataColumn2(label: const Text('Created at'), size: ColumnSize.S, onSort: _sort),
-                    DataColumn2(label: const Text('Last Login at'), size: ColumnSize.S, onSort: _sort),
-                    DataColumn2(label: const Text('Datawallet version'), onSort: _sort),
-                    DataColumn2(label: const Text('Identity Version'), onSort: _sort),
-                  ],
-                ),
-              ),
+              Expanded(child: IdentitiesDataTable(dataSource: _dataSource)),
             ],
           ),
         ),
       ),
     );
-  }
-
-  void _setRowsPerPage(int? newValue) {
-    _rowsPerPage = newValue ?? _rowsPerPage;
-    _dataSource.refreshDatasource();
-  }
-
-  void _sort(int columnIndex, bool ascending) {
-    setState(() {
-      _sortColumnIndex = columnIndex;
-      _sortColumnAscending = ascending;
-    });
-    _dataSource
-      ..sort(sortColumnIndex: _sortColumnIndex, sortColumnAscending: _sortColumnAscending)
-      ..refreshDatasource();
   }
 }
